feat: complete object tracking based on HSV values
Successfully implemented and completed the object tracking program. Removed the test folder `testFindContours_cap_hsv` as it is no longer needed. All functionality has been thoroughly tested, and everything is working as expected.

BREAKING CHANGE: None
